Removes UtilsClasses
